Refactor code to enhance clean code
