--- src/components/ButtonSkeleton.js
@@ -1,24 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-
-class Button extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    const { buttons } = this.props;
-    return (
-      <div className={buttons.length < 6 ? 'operators' : 'digits'}>
-        { buttons.map((item) => <button type="button" key={item}>{ item }</button>) }
-      </div>
-    );
-  }
-}
-
-Button.propTypes = {
-  buttons: PropTypes.string.isRequired,
-};
-
-export default Button;
